conda_requirements.txt
# This file may be used to create an environment using:
# $ conda create --name <env> --file <this file>
# platform: linux-64
_libgcc_mutex=0.1=main
_openmp_mutex=4.5=1_gnu
absl-py=0.15.0=pyhd3eb1b0_0
aiohttp=3.8.1=py38h7f8727e_0
aiosignal=1.2.0=pyhd3eb1b0_0
alabaster=0.7.12=py_0
astroid=2.9.3=py38h578d9bd_0
async-timeout=4.0.1=pyhd3eb1b0_0
attrs=21.4.0=pyhd3eb1b0_0
babel=2.9.1=pyh44b312d_0
blas=1.0=mkl
blinker=1.4=py38h06a4308_0
bottleneck=1.3.2=py38heb32a55_1
box2d=2.3.10=pypi_0
brotli=1.0.9=he6710b0_2
brotlipy=0.7.0=py38h27cfd23_1003
bzip2=1.0.8=h7f98852_4
c-ares=1.18.1=h7f8727e_0
ca-certificates=2021.10.26=h06a4308_2
cachetools=4.2.2=pyhd3eb1b0_0
certifi=2021.10.8=py38h06a4308_2
cffi=1.15.0=py38hd667e15_1
charset-normalizer=2.0.4=pyhd3eb1b0_0
click=8.0.3=pyhd3eb1b0_0
cloudpickle=2.0.0=pyhd8ed1ab_0
colorama=0.4.4=pyh9f0ad1d_0
conda=4.11.0=py38h06a4308_0
conda-content-trust=0.1.1=pyhd3eb1b0_0
conda-package-handling=1.7.3=py38h27cfd23_1
cryptography=3.4.8=py38hd23ed53_0
cudatoolkit=11.1.1=h6406543_8
cycler=0.11.0=pyhd3eb1b0_0
dataclasses=0.8=pyh6d0b6a4_7
dbus=1.13.18=hb2f20db_0
docutils=0.16=py38h578d9bd_3
expat=2.4.4=h295c915_0
ffmpeg=4.3=hf484d3e_0
fontconfig=2.13.1=h6c09931_0
fonttools=4.25.0=pyhd3eb1b0_0
freetype=2.10.4=h0708190_1
frozenlist=1.2.0=py38h7f8727e_0
gitdb=4.0.7=pyhd3eb1b0_0
gitpython=3.1.18=pyhd3eb1b0_1
glib=2.69.1=h4ff587b_1
gmp=6.2.1=h58526e2_0
gnutls=3.6.13=h85f3911_1
google-auth=1.33.0=pyhd3eb1b0_0
google-auth-oauthlib=0.4.1=py_2
grpcio=1.42.0=py38hce63b2e_0
gst-plugins-base=1.14.0=h8213a91_2
gstreamer=1.14.0=h28cd5cc_2
gym=0.21.0=py38he5a9106_2
icu=58.2=he6710b0_3
idna=3.3=pyhd3eb1b0_0
imagesize=1.3.0=pyhd8ed1ab_0
importlib-metadata=4.11.1=py38h578d9bd_0
intel-openmp=2022.0.1=h06a4308_3633
isort=5.10.1=pyhd8ed1ab_0
jinja2=3.0.3=pyhd8ed1ab_0
jpeg=9b=h024ee3a_2
kiwisolver=1.3.2=py38h295c915_0
lame=3.100=h7f98852_1001
lazy-object-proxy=1.6.0=py38h27cfd23_0
ld_impl_linux-64=2.35.1=h7274673_9
libffi=3.3=he6710b0_2
libgcc-ng=9.3.0=h5101ec6_17
libgomp=9.3.0=h5101ec6_17
libiconv=1.16=h516909a_0
libpng=1.6.37=h21135ba_2
libprotobuf=3.19.1=h4ff587b_0
libstdcxx-ng=9.3.0=hd4cf53a_17
libtiff=4.1.0=h2733197_1
libuuid=1.0.3=h7f8727e_2
libuv=1.42.0=h7f98852_0
libxcb=1.14=h7b6447c_0
libxml2=2.9.12=h03d6c58_0
lz4-c=1.9.3=h9c3ff4c_1
markdown=3.3.4=py38h06a4308_0
markupsafe=2.0.1=py38h497a2fe_0
matplotlib=3.4.3=py38h06a4308_0
matplotlib-base=3.4.3=py38hbbc1b5f_0
mccabe=0.6.1=py_1
mkl=2020.2=256
mkl-service=2.3.0=py38h1e0a361_2
mkl_fft=1.2.0=py38hab2c0dc_1
mkl_random=1.2.0=py38hc5bc63f_1
multidict=5.2.0=py38h7f8727e_2
munkres=1.1.4=py_0
ncurses=6.3=h7f8727e_2
nettle=3.6=he412f7d_0
ninja=1.10.2=h4bd325d_0
numexpr=2.7.3=py38hb2eb853_0
numpy=1.19.2=py38h54aff64_0
numpy-base=1.19.2=py38hfa32c7d_0
oauthlib=3.2.0=pyhd3eb1b0_0
olefile=0.46=pyh9f0ad1d_1
openh264=2.1.1=h780b84a_0
openssl=1.1.1m=h7f8727e_0
packaging=21.3=pyhd8ed1ab_0
pandas=1.4.1=py38h295c915_0
pcre=8.45=h295c915_0
pillow=7.1.2=py38hb39fc2d_0
pip=21.2.4=py38h06a4308_0
platformdirs=2.5.0=pyhd8ed1ab_0
protobuf=3.19.1=py38h295c915_0
pyasn1=0.4.8=pyhd3eb1b0_0
pyasn1-modules=0.2.8=py_0
pycosat=0.6.3=py38h7b6447c_1
pycparser=2.21=pyhd3eb1b0_0
pygments=2.11.2=pyhd8ed1ab_0
pyjwt=2.1.0=py38h06a4308_0
pylint=2.12.2=pyhd8ed1ab_0
pynput=1.7.3=py38h578d9bd_0
pyopenssl=21.0.0=pyhd3eb1b0_1
pyparsing=3.0.4=pyhd3eb1b0_0
pyqt=5.9.2=py38h05f1152_4
pysocks=1.7.1=py38h06a4308_0
python=3.8.12=h12debd9_0
python-dateutil=2.8.2=pyhd3eb1b0_0
python-xlib=0.28=pyh9f0ad1d_0
python_abi=3.8=2_cp38
pytorch=1.8.0=py3.8_cuda11.1_cudnn8.0.5_0
pytz=2021.3=pyhd3eb1b0_0
pyyaml=6.0=pypi_0
qt=5.9.7=h5867ecd_1
readline=8.1.2=h7f8727e_1
requests=2.27.1=pyhd3eb1b0_0
requests-oauthlib=1.3.0=py_0
rsa=4.7.2=pyhd3eb1b0_1
ruamel_yaml=0.15.100=py38h27cfd23_0
setuptools=58.0.4=py38h06a4308_0
setuptools-lint=0.6.0=pyh9f0ad1d_0
sip=4.19.13=py38h295c915_0
six=1.16.0=pyhd3eb1b0_0
smmap=4.0.0=pyhd3eb1b0_0
snowballstemmer=2.2.0=pyhd8ed1ab_0
sphinx=4.4.0=pyh6c4a22f_1
sphinxcontrib-applehelp=1.0.2=py_0
sphinxcontrib-devhelp=1.0.2=py_0
sphinxcontrib-htmlhelp=2.0.0=pyhd8ed1ab_0
sphinxcontrib-jsmath=1.0.1=py_0
sphinxcontrib-qthelp=1.0.3=py_0
sphinxcontrib-serializinghtml=1.1.5=pyhd8ed1ab_1
sqlite=3.37.0=hc218d9a_0
swig=4.0.2=h295c915_4
tensorboard=2.6.0=py_1
tensorboard-data-server=0.6.0=py38hca6d32c_0
tensorboard-plugin-wit=1.6.0=py_0
tk=8.6.11=h1ccaba5_0
toml=0.10.2=pyhd8ed1ab_0
torchaudio=0.8.0=py38
torchvision=0.9.0=py38_cu111
tornado=6.1=py38h27cfd23_0
tqdm=4.62.3=pyhd3eb1b0_1
typing-extensions=3.10.0.2=hd3eb1b0_0
typing_extensions=3.10.0.2=pyh06a4308_0
urllib3=1.26.7=pyhd3eb1b0_0
werkzeug=2.0.2=pyhd3eb1b0_0
wheel=0.37.1=pyhd3eb1b0_0
wrapt=1.11.2=py38h1e0a361_1
xz=5.2.5=h7b6447c_0
yaml=0.2.5=h7b6447c_0
yarl=1.6.3=py38h27cfd23_0
zipp=3.7.0=pyhd8ed1ab_1
zlib=1.2.11=h7f8727e_4
zstd=1.4.9=ha95c52a_0
